+++ protocol/trojan/packet.go
@@ -29,6 +29,10 @@ func (i *TrojanPacketSession) ReadPacket() (*protocol.Request, []byte, error) {
 		return req, nil, common.NewError("Failed to read length")
 	}
 	length := binary.BigEndian.Uint16(lengthBuf[:])
+
+	crlf := [2]byte{}
+	io.ReadFull(i.conn, crlf[:])
+
 	packet := make([]byte, length)
 	_, err = io.ReadFull(i.conn, packet)
 	if err != nil {
@@ -38,13 +42,19 @@ func (i *TrojanPacketSession) ReadPacket() (*protocol.Request, []byte, error) {
 }
 
 func (i *TrojanPacketSession) WritePacket(req *protocol.Request, packet []byte) (int, error) {
-	buf := bytes.NewBuffer(make([]byte, 0, 512))
+	buf := bytes.NewBuffer(make([]byte, 0, len(packet)+32))
 	common.Must(req.Address.Unmarshal(buf))
+
 	length := len(packet)
 	lengthBuf := [2]byte{}
 	binary.BigEndian.PutUint16(lengthBuf[:], uint16(length))
 	buf.Write(lengthBuf[:])
+
+	crlf := [2]byte{0x0d, 0x0a}
+	buf.Write(crlf[:])
+
 	buf.Write(packet)
+
 	return i.conn.Write(buf.Bytes())
 }
